requirements.txt
py-clob-client>=0.34.0
python-dotenv>=1.0.0
rich>=13.7.0
httpx>=0.27.0
